moved image files into image directory within src and updated imports
