+++ front-end/src/images/front-page-background.png
