+++ front-end/src/images/speakeasyLogo.PNG
